backend/requirements.txt
langchain<0.3
langchain-openai
fastapi
uvicorn
python-dotenv
pydantic
typing-extensions
psycopg[binary]
sqlalchemy
langchain-community
